+++ CHANGELOG.md
@@ -22,6 +22,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `ewoksorange.engine.OrangeWorkflowEngine` can be instantiated without importing Qt.
 
+## Changed
+
+- `OWEwoksBaseWidget.get_dynamic_input`: return the actual value, not the wrapper `Variable`.
+
 ## [2.1.0] - 2025-10-23
 
 ### Added