+++ CHANGELOG.md
@@ -24,7 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Changed
 
-- `OWEwoksBaseWidget.get_dynamic_input`: return the actual value, not the wrapper `Variable`.
+- `OWEwoksBaseWidget.get_dynamic_input_value`: return the actual value, not the wrapper `Variable`.
 
 ## [2.1.0] - 2025-10-23